+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - PR #1047 Updated select tests to use new dataset list that includes asymmetric directed graph
 - PR #1090 Add experimental Leiden function
 - PR #1077 Updated/added copyright notices, added copyright CI check from cuml
+- PR #1100 Add support for new build process (Project Flash)
 
 ## Improvements
 - PR #898 Add Edge Betweenness Centrality, and endpoints to BC